update flattened field
