+++ modules/lang-painless/src/main/resources/org/elasticsearch/painless/org.elasticsearch.script.fields.txt
@@ -92,7 +92,7 @@ class org.elasticsearch.script.field.DateNanosField @dynamic_type {
   ZonedDateTime get(int, ZonedDateTime)
 }
 
-class org.elasticsearch.script.field.KeywordField @dynamic_type {
+class org.elasticsearch.script.field.StringField @dynamic_type {
   String get(String)
   String get(int, String)
 }
@@ -102,7 +102,7 @@ class org.elasticsearch.script.field.KeywordDocValuesField @dynamic_type {
 }
 
 # subclass of AbstractKeywordDocValuesField
-class org.elasticsearch.script.field.FlattenedDocValuesField @dynamic_type {
+class org.elasticsearch.script.field.FlattenedField @dynamic_type {
 }
 
 class org.elasticsearch.script.field.GeoPointDocValuesField @dynamic_type {
